Possible BREAKING CHANGES:
Conductors have been refined based on forum discussion. I have introduced a generic IParent interface into the framework. IConductor inherits from this. As a result, instead of IConductor having a GetConductedItems method, it inherits GetChildren from IParent. Additionally, ActiveItem was removed from IConductor since not all conductors have an active item. I have added two new interfaces to address this. IHaveActiveItem which declares a single ActiveItem property and IConductActiveItem which inherits both IConductor and IHaveActiveItem. Finally, IConductor.CloseItem has been replaced by IConductor.DeactivateItem(item, close) which is more consistent with the interface and makes for more discoverable behavior. In order to keep code based on CloseItem working, I have created an extension method with the same name which retains the original behavior. Some internal implementation details of conductors have changed as a result of this and all dependent code and samples have been updated to use the new interfaces. One special thing to be aware of: If you have an Message.Attach or action conventions that were using CloseItem, that will need to be altered to use DeactivateItem.

Additionally, IChild and IHaveDisplayName have been moved out of the Screens namespace and into the ApplicationModel namespace. That is where the new IParent interface lives as well.